requirements.txt
fastapi
uvicorn[standard]
celery
redis
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
